backend/requirements.txt
flask
flask-cors
opencv-python-headless
numpy
pillow
python-dotenv
google-genai